+++ src/Concerto/PanelBundle/Resources/public/angularjs/app/concerto_panel/js/controllers/data_table_controller.js
@@ -565,8 +565,6 @@ function DataTableController($scope, $uibModal, $http, $filter, $timeout, $state
             templateUrl: Paths.DIALOG_TEMPLATE_ROOT + "import_csv_dialog.html",
             controller: DataTableImportCsvController,
             size: "lg",
-            backdrop: 'static',
-            keyboard: false,
             resolve: {
                 object: function () {
                     return $scope.object;